Refactor tag_album.py into smaller modules
This commit refactors the main `tag_album.py` script by splitting its functionalities into several smaller modules located in the new `tag_album_utils` directory.

The following modules were created:
- `tag_album_utils/display.py`: Contains functions for displaying ID3 tags in the console and as an HTML report.
- `tag_album_utils/set_tags.py`: Contains functions for setting various ID3 tags (artist, album, genre, rating, cover art, year).
- `tag_album_utils/fetch_metadata.py`: Contains the function for fetching metadata from MusicBrainz.
- `tag_album_utils/utils.py`: Contains utility functions, currently `write_log`.

The main `tag_album.py` script has been updated to import and use these modules. The test suite in `test/test_tag_album.py` has also been updated to reflect these changes, and all tests are passing.

This refactoring improves the organization and maintainability of the codebase.
