+++ tag_album_utils/__init__.py
@@ -1,0 +1,1 @@
+# This file makes tag_album_utils a Python package
